Creating interface and utils files & adding error messges to require fields
