+++ src/components/Screen/CourseEditorScreen/CourseOutlineQuiz/DraggableQuestionItem.tsx
@@ -4,7 +4,7 @@ import cc from 'classcat';
 
 import { QuizQuestion } from '../../../../walkme/data/courseBuild/quiz/question';
 
-import { QuizScreenType } from '../QuizEditForm';
+import { QuizScreenType } from '../QuizEditForm/interface';
 
 import QuestionItem from './QuestionItem';
