+++ src/components/Screen/CourseEditorScreen/QuizEditForm/interface.ts
@@ -1,0 +1,6 @@
+export enum QuizScreenType {
+  WelcomeScreen = 'welcome-screen',
+  FailScreen = 'fail-screen',
+  SuccessScreen = 'success-screen',
+  QuestionScreen = 'question-screen',
+}
